회원가입 로그인 리팩토링
