+++ homework/hyunwoo/backend/.env.example
@@ -1,0 +1,1 @@
+JWT_SECRET_KEY="이곳에 JWT SECRET KEY를 넣어주세요"
