+++ homework/hyunwoo/backend/db/users.js
@@ -1,0 +1,9 @@
+export default [
+    {
+      id: 1,
+      email: "hyunwoo@example.com",
+      password: "1234",
+      role: "student",
+      name: "이현우",
+    },
+  ];
