+++ homework/hyunwoo/backend/middlewares/errorHandler.middleware.js
@@ -1,0 +1,3 @@
+export default function errorMiddleware(err, req, res, next) {
+    res.status(500).json({ message: "Internal Server Error" });
+  }
