+++ homework/hyunwoo/backend/package.json
@@ -6,6 +6,9 @@
   "type": "module",
   "dependencies": {
     "cors": "^2.8.5",
+    "dot": "^1.1.3",
+    "dotenv": "^16.4.5",
+    "env": "^0.0.2",
     "express": "^4.19.2",
     "jsonwebtoken": "^9.0.2"
   }
